Añadida funcion para la gestion de imagenes de los usuarios e importar y exportar con CSV los coches

+++ CarRegistry/src/main/java/com/msj/CarRegistry/repository/CarRepository.java
@@ -3,6 +3,6 @@ package com.msj.CarRegistry.repository;
 import com.msj.CarRegistry.entity.CarEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarRepository extends JpaRepository<CarEntity,Integer> {
+public interface CarRepository extends JpaRepository<CarEntity,Long> {
 
 }
